items/encoding/steganography/pyproject.toml
[tool.poetry]
name = "steganography"
version = "0.1.0"
description = ""
authors = ["JKearnsl <boyarshinovn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
